+++ lib/spring/application.rb
@@ -157,6 +157,8 @@ module Spring
         IGNORE_SIGNALS.each { |sig| trap(sig, "DEFAULT") }
         trap("TERM", "DEFAULT")
 
+        puts "Running via Spring preloader in process #{Process.pid}" unless Spring.quiet
+
         ARGV.replace(args)
         $0 = command.exec_name