+++ lib/spring/client/run.rb
@@ -1,4 +1,3 @@
-require "spring/commands"
 require "rbconfig"
 require "socket"
 require "bundler"
@@ -138,8 +137,6 @@ ERROR
 
         if pid && !pid.empty?
           log "got pid: #{pid}"
-
-          puts "Running via Spring preloader in process #{pid}" unless Spring.quiet
 
           forward_signals(pid.to_i)
           status = application.read.to_i